# resources/openapi.yml
swagger: "2.0"
info:
  title: BSC API
  description: BSC API
  version: 0.0.1
host: chekov.basecondition.bsc
basePath: /api/v1
...
schemes:
  - https
paths:
...
  /message/{id}:
    post:
      tags:
        - message
      summary: Reply to a conversation
      description: Reply to a conversation ({id} = conversation (root) id)
      operationId: sendReply
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          format: "int64"
      produces:
        - application/json
      responses:
        500:
          $ref: '#/responses/500'
        403:
          $ref: '#/responses/403'
        401:
          $ref: '#/responses/401'
        201:
          description: Reply sent successfully (returns new reply ID)
          schema:
            $ref: '#/definitions/messageSend'
      security:
        - oAuth2: []
    get:
      tags:
        - message
      summary: Get conversation and all replies
      description: Get conversation and all replies
      operationId: getConversation
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          format: "int64"
      produces:
        - application/json
      responses:
        500:
          $ref: '#/responses/500'
        401:
          $ref: '#/responses/401'
        200:
          description: conversation
          schema:
            $ref: '#/definitions/messageGet'
      security:
        - oAuth2: []
...
securityDefinitions:
  basicAuth:
    type: basic
  apiKey:
    type: apiKey
    in: header
    name: X-API-Key
  oAuth2:
    type: oauth2
    flow: accessCode
    authorizationUrl: https://chekov.basecondition.bsc/api/v1/token
    tokenUrl: https://chekov.basecondition.bsc/api/v1/token
    scopes:
      read: Grants read access
      write: Grants write access
      admin: Grants read and write access to administrative information
responses:
  500:
    description: internal_server_error
    schema:
      $ref: '#/definitions/error'
...
  403:
    description: forbidden
    schema:
      $ref: '#/definitions/error'
  401:
    description: invalid_token
    schema:
      $ref: '#/definitions/error'
...
definitions:
  error:
    type: object
    properties:
      error:
        type: string
      error_description:
        type: string
...
  messageSend:
    type: object
    properties:
      content:
        type: string
  messageGet:
    type: object
    properties:
      id:
        type: integer
      read:
        type: boolean
        description: flag if conversation root message and every reply from BTU side was read by client
      closed:
        type: boolean
        description: conversation was closed by BTU
      content:
        type: string
      createdate:
        type: string
      createtimestamp:
        type: integer
        description: unix timestamp (seconds)
      createdby:
        type: string
        enum:
          - API
          - BACKEND_USER
        description: API (user) or BACKEND_USER (BTU)
      replies:
        type: integer
        description: sum of all replies
      latest_reply:
        type: string
        description: date of latest reply
      latest_reply_read:
        type: boolean
        description: latest reply was read
      hash:
        type: string
        description: sha1 hash of whole conversation to quickly identy changes
      conversation:
        type: array
        description: all replies ordered by create date descending
        items:
          properties:
            id:
              type: integer
              description: id of the reply message
            read:
              type: boolean
            content:
              type: string
            createdate:
              type: string
            createtimestamp:
              type: integer
              description: unix timestamp (seconds)
            createdby:
              type: string
              enum:
                - API
                - BACKEND_USER
              description: API (user) or BACKEND_USER (BTU)
...